Refactor the Lookup table fields into a common component
