+++ graylog2-web-interface/src/components/event-definitions/event-definition-form/field-value-providers/LookupTableFieldValueProviderForm.jsx
@@ -17,19 +17,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import cloneDeep from 'lodash/cloneDeep';
-import memoize from 'lodash/memoize';
 
-import { Col, ControlLabel, FormGroup, HelpBlock, Row } from 'components/bootstrap';
-import { Select } from 'components/common';
-import * as FormsUtils from 'util/FormsUtils';
-import { naturalSortIgnoreCase } from 'util/SortUtils';
+import { LookupTableFields } from 'components/common';
 
 class LookupTableFieldValueProviderForm extends React.Component {
   static propTypes = {
-    allFieldTypes: PropTypes.array.isRequired,
     config: PropTypes.object.isRequired,
     validation: PropTypes.object.isRequired,
-    lookupTables: PropTypes.array.isRequired,
     onChange: PropTypes.func.isRequired,
   };
 
@@ -45,16 +39,6 @@ class LookupTableFieldValueProviderForm extends React.Component {
     'key_field',
   ];
 
-  formatMessageFields = memoize(
-    (fieldTypes) => fieldTypes
-      .sort((ftA, ftB) => naturalSortIgnoreCase(ftA.name, ftB.name))
-      .map((fieldType) => ({
-        label: `${fieldType.name} – ${fieldType.value.type.type}`,
-        value: fieldType.name,
-      })),
-    (fieldTypes) => fieldTypes.map((ft) => ft.name).join('-'),
-  );
-
   propagateChanges = (key, value) => {
     const { config, onChange } = this.props;
     const nextProviders = cloneDeep(config.providers);
@@ -64,58 +48,20 @@ class LookupTableFieldValueProviderForm extends React.Component {
     onChange({ ...config, providers: nextProviders });
   };
 
-  handleChange = (event) => {
-    const { name } = event.target;
-    const value = FormsUtils.getValueFromInput(event.target);
-
-    this.propagateChanges(name, value);
-  };
-
   handleSelectChange = (key) => (nextLookupTable) => {
     this.propagateChanges(key, nextLookupTable);
   };
 
-  formatLookupTables = (lookupTables) => lookupTables
-    .sort((lt1, lt2) => naturalSortIgnoreCase(lt1.title, lt2.title))
-    .map((table) => ({ label: table.title, value: table.name }));
-
   render() {
-    const { allFieldTypes, config, lookupTables, validation } = this.props;
+    const { config, validation } = this.props;
     const provider = config.providers.find((p) => p.type === LookupTableFieldValueProviderForm.type);
 
     return (
-      <Row className="row-sm">
-        <Col md={7} lg={6}>
-          <FormGroup controlId="lookup-provider-table" validationState={validation.errors.table_name ? 'error' : null}>
-            <ControlLabel>Select Lookup Table</ControlLabel>
-            <Select name="event-field-table-name"
-                    placeholder="Select Lookup Table"
-                    onChange={this.handleSelectChange('table_name')}
-                    options={this.formatLookupTables(lookupTables)}
-                    value={provider.table_name}
-                    matchProp="label"
-                    required />
-            <HelpBlock>
-              {validation.errors.table_name || 'Select the Lookup Table Graylog should use to get the value.'}
-            </HelpBlock>
-          </FormGroup>
-
-          <FormGroup controlId="lookup-provider-table" validationState={validation.errors.key_field ? 'error' : null}>
-            <ControlLabel>Lookup Table Key Field</ControlLabel>
-            <Select name="lookup-provider-key"
-                    placeholder="Select Field"
-                    onChange={this.handleSelectChange('key_field')}
-                    options={this.formatMessageFields(allFieldTypes)}
-                    value={provider.key_field}
-                    matchProp="label"
-                    allowCreate
-                    required />
-            <HelpBlock>
-              {validation.errors.key_field || 'Message Field name whose value will be used as Lookup Table Key.'}
-            </HelpBlock>
-          </FormGroup>
-        </Col>
-      </Row>
+      <LookupTableFields onTableNameChange={this.handleSelectChange('table_name')}
+                         onKeyChange={this.handleSelectChange('key_field')}
+                         selectedTableName={provider.table_name}
+                         selectedKeyName={provider.key_field}
+                         validation={validation} />
     );
   }
 }
